# Conflicts:
#	robot-tests/IFS_acceptance_tests/tests/10__Project_setup/10__mandatory_documents.robot

+++ ifs-resources/src/test/java/org/innovateuk/ifs/project/resource/ProjectResourceTest.java
@@ -50,7 +50,7 @@ public class ProjectResourceTest {
         assertEquals(projectResource.getAddress(), addressResource);
         assertEquals(projectResource.getName(), name);
         assertEquals(projectResource.getDurationInMonths(), duration);
-        assertEquals(projectResource.getApplication(), applicationResource.getId());
+        assertEquals(projectResource.getApplication(), (long) applicationResource.getId());
         assertEquals(projectResource.getProjectUsers(), projectUsers);
     }
